requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# PDF Processing
PyMuPDF==1.25.3
pytesseract==0.3.13
Pillow>=11.2.0

# Language Detection
langdetect==1.0.9

# Embeddings
sentence-transformers==3.4.1

# Vector Database
qdrant-client==1.13.3

# LLM
transformers==4.48.3
torch>=2.1.0
accelerate==1.3.0

# Reranker
# (uses sentence-transformers cross-encoder)

# ServiceNow HTTP Client
httpx==0.28.1

# Logging
structlog==25.1.0
